Change how diagnostics are gathered
