+++ src/EditorFeatures/Core/Implementation/PasteTracking/PasteTrackingService.cs
@@ -3,9 +3,7 @@
 using System.Composition;
 using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 using Microsoft.CodeAnalysis.Text;
-using Microsoft.VisualStudio.Language.Intellisense;
 using Microsoft.VisualStudio.Text;
-using Microsoft.VisualStudio.Text.Editor;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.PasteTracking
@@ -15,15 +13,11 @@ namespace Microsoft.CodeAnalysis.PasteTracking
     internal class PasteTrackingService : IPasteTrackingService
     {
         private readonly IThreadingContext _threadingContext;
-        private readonly ILightBulbBroker2 _lightBulbBroker;
-        private readonly ISuggestedActionCategorySet _refactoringsCategorySet;
 
         [ImportingConstructor]
-        public PasteTrackingService(IThreadingContext threadingContext, ILightBulbBroker2 lightBulbBroker, ISuggestedActionCategoryRegistryService categoryRegistryService)
+        public PasteTrackingService(IThreadingContext threadingContext)
         {
             _threadingContext = threadingContext;
-            _lightBulbBroker = lightBulbBroker;
-            _refactoringsCategorySet = categoryRegistryService.AllRefactorings;
         }
 
         public bool TryGetPastedTextSpan(Document document, out TextSpan textSpan)
@@ -38,16 +32,12 @@ namespace Microsoft.CodeAnalysis.PasteTracking
             return textBuffer.Properties.TryGetProperty(this, out textSpan);
         }
 
-        internal void RegisterPastedTextSpan(ITextView textView, ITextBuffer textBuffer, TextSpan textSpan)
+        internal void RegisterPastedTextSpan(ITextBuffer textBuffer, TextSpan textSpan)
         {
             Contract.ThrowIfFalse(_threadingContext.HasMainThread);
 
             textBuffer.Changed += RemovePastedTestSpan;
             textBuffer.Properties.AddProperty(this, textSpan);
-
-            // Try to start the light bulb session in an unexpanded state now
-            // so that it displayed to the user sooner.
-            TryShowLightBulb(textView, textSpan);
 
             return;
 
@@ -69,19 +59,6 @@ namespace Microsoft.CodeAnalysis.PasteTracking
 
             textBuffer = text.Container.TryGetTextBuffer();
             return textBuffer != null;
-        }
-
-        private void TryShowLightBulb(ITextView textView, TextSpan textSpan)
-        {
-            // Use the current caret positon as the trigger point and span
-            var caretPosition = textView.Caret.Position.BufferPosition;
-            var triggerPoint = textView.TextBuffer.CurrentSnapshot.CreateTrackingPoint(caretPosition, PointTrackingMode.Negative);
-            var triggerSpan = textView.TextBuffer.CurrentSnapshot.CreateTrackingSpan(caretPosition.Position, 0, SpanTrackingMode.EdgeExclusive);
-
-            // Start a suggested refactoring session. Session ends if no suggested actions are found.
-            var session = _lightBulbBroker.CreateSession(
-                _refactoringsCategorySet, textView, triggerPoint, triggerSpan, _refactoringsCategorySet, trackMouse: false);
-            session.Start();
         }
     }
 }
